refactor: change hook file names to match the rest of the application
