+++ src/components/editor-page/editor-page-content.tsx
@@ -11,13 +11,13 @@ import { updateNoteTitleByFirstHeading } from '../../redux/note-details/methods'
 import { MotdModal } from '../common/motd-modal/motd-modal'
 import { AppBar, AppBarMode } from './app-bar/app-bar'
 import { EditorMode } from './app-bar/editor-view-mode'
-import { useViewModeShortcuts } from './hooks/useViewModeShortcuts'
+import { useViewModeShortcuts } from './hooks/use-view-mode-shortcuts'
 import { Sidebar } from './sidebar/sidebar'
 import { Splitter } from './splitter/splitter'
 import type { DualScrollState, ScrollState } from './synced-scroll/scroll-props'
-import { useEditorModeFromUrl } from './hooks/useEditorModeFromUrl'
+import { useEditorModeFromUrl } from './hooks/use-editor-mode-from-url'
 import { UiNotifications } from '../notifications/ui-notifications'
-import { useUpdateLocalHistoryEntry } from './hooks/useUpdateLocalHistoryEntry'
+import { useUpdateLocalHistoryEntry } from './hooks/use-update-local-history-entry'
 import { useApplicationState } from '../../hooks/common/use-application-state'
 import { EditorDocumentRenderer } from './editor-document-renderer/editor-document-renderer'
 import { Logger } from '../../utils/logger'
