+++ src/components/editor-page/hooks/use-view-mode-shortcuts.ts
